Cast -1 values to unsigned to quiet the compiler.
Marking the magic values to indicate we really did mean to type them.

+++ src/cairo-pattern.c
@@ -1303,7 +1303,8 @@ _cairo_pattern_acquire_surface_for_surface (cairo_surface_pattern_t   *pattern,
 	/* If we're repeating, we just play it safe and clone the entire surface. */
 	/* If requested width and height are -1, clone the entire surface.
 	 * This is relied on in the svg backend. */
-	if (attr->extend == CAIRO_EXTEND_REPEAT || (width == -1 && height == -1)) {
+	if (attr->extend == CAIRO_EXTEND_REPEAT ||
+	    (width == (unsigned int) -1 && height == (unsigned int) -1)) {
 	    x = extents.x;
 	    y = extents.y;
 	    width = extents.width;
